+++ Magnus.Application/Services/Interfaces/IProductStockAppService.cs
@@ -1,4 +1,3 @@
-using System.Linq.Expressions;
 using Magnus.Application.Dtos.Responses;
 using Magnus.Core.Entities;
 
@@ -6,7 +5,9 @@ namespace Magnus.Application.Services.Interfaces;
 
 public interface IProductStockAppService
 {
-    Task<IEnumerable<ProductStockResponse>> GetProductStocksByFilterAsync(Guid productId, int warehouseId, bool all, CancellationToken cancellationToken);
-    Task<decimal> GetBalanceProductStocksAsync(Guid productId, int warehouseId, CancellationToken cancellationToken);
+    Task<IEnumerable<ProductStockResponse>> GetProductStocksByFilterAsync(Guid productId, int warehouseId, bool all,
+        CancellationToken cancellationToken);
+
+    Task<int> GetBalanceProductStocksAsync(Guid productId, int warehouseId, CancellationToken cancellationToken);
     Task CreateProductStockMovementAsync(ProductStock productStock, CancellationToken cancellationToken);
 }
